Allow a special case for .gitignore and .npmignore files
npm doesn't have a mechanism to package `.gitignore and` `.npmignore` files
as they are used to modify the npm package itself.

Add a special file name prefix `.dot.` that will be converted to `.whatever`
at templating time so any special . files to npm can be included in a template
set.

    template/files/.dot.gitignore -> .gitignore
    template/files/.dot.npmignore -> .npmignore